Add roles restrictions for room editing, deleting

+++ server/db/managers/rooms.js
@@ -7,66 +7,45 @@ const MAX_PLAYERS_IN_THE_ROOM = 6;
 const { getObjEnabledFields } = require('../../helpers/validators');
 const CREATE_FIELDS = ['name', 'state'];
 const READ_FIELDS = ['id', 'name', 'state', 'created_at', 'updated_at'];
-/////////////////////////////////////////////////////////////////////////////
+
+function makeAnswerWithError(detail) {
+  log.warn(detail);
+  return { success: false, error: { detail } };
+}
+
 const newPlayer = async (id, userId) => {
   const room = await Rooms.query()
     .findById(id)
     .eager('users(onlyIds)');
   if (!room) {
-    log.error(`The room with id = ${id} not found.`);
-    return {
-      success: false,
-      error: { detail: `The room with id = ${id} not found.` }
-    };
+    return makeAnswerWithError(`The room with id = ${id} not found.`);
   }
   if (room.users && room.users.length >= MAX_PLAYERS_IN_THE_ROOM) {
-    log.error(`The room with id = ${id} already filled.`);
-    return {
-      success: false,
-      error: { detail: `The room with id = ${id} already filled.` }
-    };
+    return makeAnswerWithError(`The room with id = ${id} already filled.`);
   }
   if (room.users.find(user => user.id === userId)) {
-    log.error(`The user ${userId} alredy in room ${id}.`);
-    return {
-      success: false,
-      error: { detail: `The user ${userId} alredy in room ${id}.` }
-    };
+    return makeAnswerWithError(`The user ${userId} alredy in room ${id}.`);
   }
-
   const user = await Users.query().findById(userId);
   if (!user) {
-    return {
-      success: false,
-      error: { detail: `The user with id = ${userId} not found.` }
-    };
+    return makeAnswerWithError(`The user with id = ${userId} not found.`);
   }
-
-  let result = await Users_rooms.query().insert({
+  await Users_rooms.query().insert({
     room_id: id,
     user_id: userId
   });
-
   return read(id);
 };
-/////////////////////////////////////////////////////////////////////////////
+
 const create = async ({ name, state }, userId) => {
-  const user = await Users.query().findById(userId);
-  if (!user) {
-    return {
-      success: false,
-      error: { detail: `The user - room creator not found.` }
-    };
-  }
   const rooms = await findByName(name);
   if (rooms.length) {
-    log.error(`The room name should be unique.`);
-    return {
-      success: false,
-      error: { detail: `The room name should be unique.` }
-    };
+    return makeAnswerWithError(`The room name should be unique.`);
   }
-
+  const user = await Users.query().findById(userId);
+  if (!user) {
+    return makeAnswerWithError(`The user - room creator not found.`);
+  }
   const result = await user.$relatedQuery('rooms').insert({ name, state });
   const { id, email } = user;
   result.users = [{ id, email }];
@@ -77,37 +56,31 @@ const read = async (id, userId) => {
   let result = id ? Rooms.query().findById(id) : Rooms.query();
   result = await result
     .eager('users(withoutPW, orderByName)')
-    // .eager('users()')
     .select(...READ_FIELDS);
-  // if (result.length) {
-  //   result.forEach((room, i) => {
-  //     if (room.users.length) {
-  //       result[i].users = room.users.map(user => user.email).join(', ');
-  //     }
-  //   });
-  // }
-  // log.info('Manager read ROOMS %O', result);
   return { success: true, result };
 };
 
 const update = async (record, userId) => {
   const roomReadAnswer = await read(record.id, userId);
   const notUpdatedRoom = roomReadAnswer.result;
-  if (!notUpdatedRoom) {
-    return {
-      success: false,
-      error: { detail: `The room ${record.id} not found.` }
-    };
-  }
+  if (!notUpdatedRoom)
+    return makeAnswerWithError(`The room ${record.id} not found.`);
   if (record.name) {
     const rooms = await findByName(record.name);
     if (rooms.length) {
-      return {
-        success: false,
-        error: { detail: `The room name should be unique.` }
-      };
+      return makeAnswerWithError(`The room name should be unique.`);
     }
   }
+
+  const user = await Users.query().findById(userId);
+  if (
+    user.role !== 'su' &&
+    user.role !== 'admin' &&
+    !notUpdatedRoom.users.some(({ id }) => id === userId)
+  ) {
+    return makeAnswerWithError(`Only entered into room users could rename it.`);
+  }
+
   const result = await Rooms.query()
     .patchAndFetchById(record.id, {
       ...notUpdatedRoom,
@@ -118,6 +91,13 @@ const update = async (record, userId) => {
 };
 
 const del = async (id, userId) => {
+  const user = await Users.query().findById(userId);
+  if (!user) {
+    return makeAnswerWithError(`The user with id = ${userId} not found.`);
+  }
+  if (user.role !== 'su' && user.role !== 'admin') {
+    return makeAnswerWithError(`You haven't rights for room deleting.`);
+  }
   await Rooms.query().deleteById(id);
   return { success: true, id };
 };
